* finished implementation of Frelon::SerialLine::splitCmd * defined Frelon::SerialLine timeouts as double constants * pass through negative (special) values in Espia sec-to/from-usec conv. * inherit EspiaSerialLine also from Espia for common consts and methods * removed unsigned long warning in EspiaAcq::getAcqFramePtr
git-svn-id: https://scm.blissgarden.org/svn/lima/trunk@72 45c4679d-1946-429d-baad-37912b19538b

+++ camera/common/espia/src/EspiaAcq.cpp
@@ -173,7 +173,7 @@ void *EspiaAcq::getBufferFramePtr(int buffer_nb, int frame_nb)
 
 void *EspiaAcq::getAcqFramePtr(int acq_frame_nb)
 {
-	int buffer_nb = ESPIA_ACQ_ANY;
+	unsigned long buffer_nb = ESPIA_ACQ_ANY;
 	void *ptr;
 	CHECK_CALL(espia_frame_address(m_dev, buffer_nb, acq_frame_nb, &ptr));
 	return ptr;
